@@ -2,7 +2,6 @@ package shoveler
 
 import (
 	"errors"
-	"io/ioutil"
 	"math/rand"
 	"net/url"
 	"os"
@@ -139,7 +138,7 @@ func readMsg(messagesQueue chan<- []byte, queue *ConfirmationQueue) {
 func readToken(tokenLocation string) (string, error) {
 	// Get the token password
 	// Read in the token file
-	tokenContents, err := ioutil.ReadFile(tokenLocation)
+	tokenContents, err := os.ReadFile(tokenLocation)
 	if err != nil {
 		log.Errorln("Unable to read file:", tokenLocation)
 		return "", err
